finish the the function to print the map

+++ include/matrix.h
@@ -8,13 +8,13 @@
 typedef struct {
     int nRows;
     int nCols;
-    double **data; // Pointer-to-pointer for 2D array
+    int **data; // Pointer-to-pointer for 2D array
 } Matrix;
 
 // Function declarations (prototypes)
 Matrix *createMatrix(int nRows, int nCols);
 void freeMatrix(Matrix *matrix);
-void setElement(Matrix *matrix, int row, int col, double value);
-double getElement(Matrix *matrix, int row, int col);
+void setElement(Matrix *matrix, int row, int col, int value);
+int getElement(Matrix *matrix, int row, int col);
 
 #endif // MATRIX_H
